requirements.txt
numpy==1.21.3
pyserial==3.5
